spike: Import React components explicitly in front end (client side only)
Server side can follow later. It's mostly there, but the React imports will need to be chunked and/or split out in some other way. It's possible a separate `vite` config might be needed as it's not been possible to far to get both client/server bundling to work differently when contained in the same config file. Once bundled correctly the Web project config should be updated where `todo`s are currently. The `NodeJS` dependency will need to be investigated, ideally by pushing out to Azure as a new app service for testing. If it is only needed for server side support then it is possible that this can be looked at later, once the client version is stable.