feat: Implemented PRG pattern for `FinancialData` actions, using data annotations to enforce model validation
Main driver behind PRG was the necessity to bind dependent data for the view from the various API calls as well as merge in submitted form data (in the case of validation errors). Without this, the POST action would have to duplicate the API call logic unnecessarily.

PRG source courtesy:
  - https://andrewlock.net/post-redirect-get-using-tempdata-in-asp-net-core/
  - https://speednetsoftware.com/post-redirect-get-in-asp-net-razor-pages/