Doc/update progress8 banding cat (#3122)
* Banding edge cases

* draft update

* Update ks4-progress-indicator.md

* Update ks4-progress-indicator.md

* Update ks4-progress-indicator.md (#3126)

corrected SUPP

* Fix grammar and clarify Progress 8 data source details

Corrected minor grammatical errors and improved clarity in the Progress 8 data source section. Added a note regarding the exclusion of schools with invalid performance outcomes.

* Update documentation/data/sources/ks4-progress-indicator.md

Co-authored-by: murdo-hippo <murdo.moyse@hippodigital.co.uk>

* Fix link format in KS4 progress indicator documentation

Corrected the link format for key-stage-4 documentation.

* Measures folder

Created a sub-folder named measures to save doc

* update title

The folder name "measure" is vague and open for interpretation. Change it to "school-performance-measure"

---------

Co-authored-by: murdo-hippo <murdo.moyse@hippodigital.co.uk>